Move back to using Trie based keyword checking
This basically reverts the change to move to a hash based approach. The reason is simple: this commit consistently improves a benchmark from a median of 2020 to a median of 1960 over hundreds of runs.

I'll have to do some more benchmarks to check whether subtle differences can make an improvement. But for now, this should be a ~3% perf improvement.

        best   median  avg
Before: 1969 / 2019 / 2028
After:  1912 / 1966 / 1977