Refactor the regex-class-escape lexer into a jump table
This does not seem to move the perf needle either way, tested on a 5mb regex (es5) test file consisting of almost exclusively regexes. In this benchmark this function gets called some 670k times.

I think the change is easier to read and maintain.

Additionally there was some dead code which has been eliminated. Some error messages have to be changed but I'm going to do that in a next commit to consolidate the changes to the tests there, proving that these changes don't change any test outcome in this commit :)