Add benchmark comparison method. Pass on separate annexb flag in various test files
Basically the benchmarks were sometimes needing module+annexb and for legacy reasons I don't have that stuff setup since I only test for sloppy+web. This passes on the annexb flag to various parts of the benchmark/testing stuff to prepare for better handling this. There's probably no real benchmark where this flag is ever false, but oh well.

Also added a way to easily get a read on the impact of a commit. Basically you go to current master, create a build (`./t z --no-compat`), then copy the build file to `./ignore/`. Then for your changes, do `./t stable; ./t --stabled p6 --parser ignore/tenko.build.mjs` and it will run all inputs on a build of the current code and the build you created before, collecitng stats on every run, and dumping it into a file to cat for later.

Alternatively, you can specify a benchmark file (and mode) and set `--max 0` to have it pinpong between the two parsers and report deltas endlessly.

While there are still some minor tweaks to be done, this is mostly done and hopefully that's the end of benchmark meta for now.

Probably not.