Massive refactor of regex parser, fix unicode "ruby" escape edge cases
Was increasing code coverage in the lexer and one thing led to another.

Basically, I learned that `\u{..}` escapes are commonly called "ruby escapes", and that kind of stuck for me. So I've replaced the awkward naming with "ruby" where I found it. Probably missed some places, who cares.

Also learned that ruby escapes can actually be valid in web compat mode. There they are considered an atom `\u`, `{`, each hex digit is an atom, and `}`.

Unfortunately, if the escape only contained numbers, it's actually the `\u` quantified. This means we have to prevent another quantifier in the case of `\u{1}` but accept it for cases like `\u{a}`.

There are also a bunch of weird edge cases to take into account for char class ranges now, since `\u{1}` in a char class might end an open range with `\u` and open a new range with `}`. Worse yet, if the escape is illegal, the latter could be arbitrary. Oh and it still has to track the legal case when a u-flag _is_ present.

So yeah, this required a massive refactor. I've tried to synchronize the code. I've deduped the various quad/ruby escape parsers.

I've added a bunch of new tests for the ruby escape case. I've checked all the tests to either only be error message changing, or that the change was correct.

Let's hope I didn't regress. Or have to touch this code again, any time soon.