Add parser option errorCodeFrame and change parser option `fullErrorContext` to `truncCodeFrame`
This is a small backwards breaking API change where default stays the same.

This allows us to disable printing the code frame in the error message. It's useful for me when running the test case minifier, since that code relies on the output message staying the same. With a code frame that would always print the input code and its line numbers, so it wouldn't be able to do anything.