Support `nodeRange` option, allow setting some options in test cases
Added some tentative tests for the `ranges` option. Also documented the `ranges` option in the README since it was missing.

Added support and a tentative test for the `nodeRange` option. This adds a `range` property to each node (hopefully I got them all) with the start and end. This should allow you to `input.slice(start, end)` to get the text that the node covers. Apparently this was needed for eslint (Fixes #11).

Added test support and a simple case for running without location tracking. Only had to adjust some assertions :)