Fix building for zsh
The file t.sh is using ";&" for cases and that seems to be compatible
with bash but not with zsh. As Apple has replaced bash by zsh back in
2019 the script is not working out of the box in current Mac computers.

I have fixed the issue by removing that notation and so the current
script should work without issues in both bash and zsh.